feat: write raw byte.
This is a simple passthrough to the method on the controller.  Most users
probably won't want it, but when a max7219 is connected to a dot-matrix display
and used as a display with `embedded_graphics` it may make sense to be able to
address the display in the smallest possible unit.  (For instance, an efficient
display updating algorithm can keep the previous state in a framebuffer and only
send the commands needed to update those bytes which have changed.)